strip csv headers
