+++ example/pretrust2.csv
@@ -1,4 +1,3 @@
-i,v
 alice,0.14285714285714285
 bob,0.14285714285714285
 2,0.14285714285714285
