+++ src/basic/util.rs
@@ -9,6 +9,19 @@ use env_logger;
 
 use log::Level;
 
+#[cfg(target_arch = "wasm32")]
+pub fn current_time_millis() -> u64 {
+    0 // todo
+}
+
+#[cfg(not(target_arch = "wasm32"))]
+pub fn current_time_millis() -> u64 {
+    // todo
+    use std::time::Instant;
+    Instant::now().elapsed().as_millis() as u64  
+}
+
+
 pub fn init_logger() {
     #[cfg(target_arch = "wasm32")]
     {
@@ -55,4 +68,22 @@ impl PeersMap {
     pub fn get_max_value(&self) -> usize {
         self.max_value
     }
+}
+
+
+pub fn strip_headers<'a>(csv_content: &'a str) -> &'a str {
+    let mut lines = csv_content.lines();
+
+    if let Some(first_line) = lines.next() {
+        let fields: Vec<&str> = first_line.split(',').collect();
+        if let Some(last_field) = fields.last() {
+            if last_field.parse::<f64>().is_err() {
+                // Return everything after the first line
+                let rest = lines.collect::<Vec<&str>>().join("\n");
+                return &csv_content[csv_content.len() - rest.len()..];
+            }
+        }
+    }
+
+    csv_content
 }
