+++ src/main.rs
@@ -26,25 +26,7 @@ fn main() {
         ::read_to_string(pretrust_csv_path)
         .expect("Failed to read pretrust CSV file");
 
-    let localtrust_csv = strip_headers(localtrust_csv);
-    let pretrust_csv = strip_headers(pretrust_csv);
-
     let result = calculate_from_csv(&localtrust_csv, &pretrust_csv).unwrap();
 
     println!("{:?}", result);
-}
-
-fn strip_headers(csv_content: String) -> String {
-    let mut lines = csv_content.lines();
-    
-    if let Some(first_line) = lines.next() {
-        let fields: Vec<&str> = first_line.split(',').collect();
-        if let Some(last_field) = fields.last() {
-            if last_field.parse::<f64>().is_err() {
-                return lines.collect::<Vec<&str>>().join("\n");
-            }
-        }
-    }
-
-    csv_content
 }
